Verified all calls.

+++ src/main/java/com/example/assignment_1/business/controller/ArtistController.java
@@ -5,17 +5,13 @@ import com.example.assignment_1.business.model.Concert;
 import com.example.assignment_1.business.model.User;
 import com.example.assignment_1.business.service.implementation.ArtistServiceImpl;
 import com.example.assignment_1.business.service.implementation.UserServiceImpl;
-import com.example.assignment_1.business.service.interfaces.ArtistService;
-import com.example.assignment_1.business.service.interfaces.UserService;
-import com.example.assignment_1.data.model.UserRole;
+import com.example.assignment_1.data.model.ArtistDB;
+import com.example.assignment_1.helper.UserRole;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.autoconfigure.quartz.QuartzTransactionManager;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -34,53 +30,57 @@ public class ArtistController {
     }
 
     @GetMapping("/getById/{artistId}")
-    public @ResponseBody Artist findById(@PathVariable Long artistId){
-        return artistService.findById(artistId);
+    public ResponseEntity<Artist> findById(@PathVariable Long artistId){
+        Artist artist = artistService.findById(artistId);
+        if(artist == null){
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        }
+
+        return new ResponseEntity<>(artist, HttpStatus.OK);
     }
 
     @PostMapping("/create")
-    public ResponseEntity create(@RequestBody Artist artist, @RequestHeader("Token") String token){
+    public ResponseEntity<Artist> create(@RequestBody Artist artist, @RequestHeader("Token") String token){
         User requestingUser = userService.findByToken(token);
 
         if (requestingUser.getRole() != UserRole.Administrator) {
-            return new ResponseEntity(HttpStatus.UNAUTHORIZED);
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
         }
 
-        artistService.save(artist);
-        return new ResponseEntity(HttpStatus.OK);
-    }
-
-    @DeleteMapping("/deleteAll")
-    public ResponseEntity deleteAll(@RequestHeader("Token") String token){
-        User requestingUser = userService.findByToken(token);
-
-        if (requestingUser.getRole() != UserRole.Administrator) {
-            return new ResponseEntity(HttpStatus.UNAUTHORIZED);
-        }
-        artistService.deleteAll();
-        return new ResponseEntity(HttpStatus.OK);
+        Artist savedArtist = new Artist(artistService.save(artist));
+        return new ResponseEntity<>(savedArtist, HttpStatus.OK);
     }
 
     @DeleteMapping("/deleteById/{artistId}")
-    public ResponseEntity deleteById(@PathVariable Long artistId, @RequestHeader("Token") String token){
+    public ResponseEntity<Artist> deleteById(@PathVariable Long artistId, @RequestHeader("Token") String token){
         User requestingUser = userService.findByToken(token);
 
         if (requestingUser.getRole() != UserRole.Administrator) {
-            return new ResponseEntity(HttpStatus.UNAUTHORIZED);
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
+        Artist artist = artistService.findById(artistId);
+        if(artist == null || artist.getId() == null){
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
         }
         artistService.deleteById(artistId);
-        return new ResponseEntity(HttpStatus.OK);
+        return new ResponseEntity<>(artist, HttpStatus.OK);
     }
 
     @PostMapping("/updateById/{artistId}")
-    public ResponseEntity update(@PathVariable Long artistId, @RequestBody Artist newValue,@RequestHeader("Token") String token){
+    public ResponseEntity<Artist> update(@PathVariable Long artistId, @RequestBody Artist newValue,@RequestHeader("Token") String token){
         User requestingUser = userService.findByToken(token);
 
         if (requestingUser.getRole() != UserRole.Administrator) {
-            return new ResponseEntity(HttpStatus.UNAUTHORIZED);
+            return new ResponseEntity<>(null,HttpStatus.UNAUTHORIZED);
         }
-        artistService.update(artistId, newValue);
-        return new ResponseEntity(HttpStatus.OK);
+
+        Artist updated =  artistService.update(artistId, newValue);
+        if(updated == null) {
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        }
+
+        return new ResponseEntity<>(updated, HttpStatus.OK);
     }
 
     @GetMapping("/getArtistConcertsById/{artistId}")
